msm: mdss: avoid skip pipe reconfiguration if pipe was freed
Current driver does not configure the pipe if the requested
data is same as the data already in the pipe.
This is a problem and can cause underruns in cases where
the pipe is freed due some error and driver tries to
setup previous configuration.
These changes make sure that if pipe was freed due some error,
driver will not skip the re-configuration of the pipe.

Change-Id: I084f367a4491244c5064cc9b55d2952d278835ce
Signed-off-by: Ingrid Gallardo <ingridg@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: placiano <placiano80@gmail.com>